requirements.txt
# QuickBot Python Dependencies

# AI providers
openai>=1.12.0
anthropic>=0.25.0
requests>=2.31.0

# Platforms
python-telegram-bot>=21.0

# Database
sqlalchemy>=2.0.0

# Configuration
pyyaml>=6.0

# Utilities
python-dotenv>=1.0.0
schedule>=1.2.0

# Optional: Advanced features
# faiss-cpu>=1.8.0  # Vector search
# tiktoken>=0.5.0   # Token counting
